Center contact icons in the footer
This commit changes the way contact icons are displayed in the footer:
instead of using <li> tags which inherit a 30px left margin, they are
now just simple <a> tags handled by flex to center them.

The new style rules are added to the `.icons` class in
`assets/css/theme.scss`.

Signed-off-by: Filippo Fontana <filippofontana1998@gmail.com>